+++ odoo/local-src/account_tax_one_vat/models/__init__.py
@@ -1,0 +1,4 @@
+from . import acccount_invoice
+from . import product
+from . import purchase_order
+from . import sale