+++ odoo/local-src/stock_delivery_note/tests/test_delivery_note.py
@@ -16,9 +16,7 @@ class TestStockDeliveryNote(SavepointCase):
         # Create a sale tax
         cls.tax = cls.env['account.tax'].create(
             {
-                'tax_group_id': cls.env.ref(
-                    'stock_delivery_note.vat_tax_group'
-                ).id,
+                'tax_group_id': cls.env.ref('specific_data.vat_tax_group').id,
                 'amount': 6,
                 'name': 'test_tax',
             }