+++ odoo/songs/upgrade/update_10_0_1_49_0.sql
@@ -1,0 +1,4 @@
+UPDATE ir_model_data
+SET module = 'specific_data'
+WHERE module = 'stock_delivery_note'
+AND name = 'vat_tax_group';